pyproject.toml
[tool.poetry]
name = "ntropy-ai"
version = "0.0.6-alpha"
description = "Unleash the power of multimodal agents"
authors = ["Hugo Le Belzic <hugolebelzic@gmail.com>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cryptography = "42.0.7"
pydantic = "2.8.2"
tabulate = "0.9.0"
Pillow = "10.4.0"

# A list of all of the optional dependencies, some of which are included in the
# below `extras`. They can be opted into by apps.
boto3 = { optional = true, version = "1.34.140" }
botocore = { optional = true, version = "1.34.140" }
openai = { optional = true, version= '1.35.10' }
pinecone-client = { optional = true, version = "4.1.2" }
pymupdf = { optional = true, version = "1.24.7" }
torch = { optional = true, version = "2.3.1" }
torchvision = { optional = true, version = "0.18.1" }
ollama = { optional = true, version = "0.2.1" }
opensearch-py = { optional = true, version = "2.6.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
providers_aws = ["boto3", "botocore", "opensearch-py"]
providers_openai = ["openai", "torch", "torchvision"] # pip install git+https://github.com/openai/CLIP.git
providers_ollama = ["ollama"]
providers_anthropic = ["anthropic"]
document_instance_pdf = ["pymupdf"]
document_instance_html = ["beautifulsoup4"]
rag_vectorstore_pinecone = ["pinecone-client"]
